Return empty string instead of null in userAgent()
Fixes #5.
Previously, null values from the underlying Request::get() would be passed as function results, which raised an exception, since this function is supposed to return string values only.

I chose the (imho) sensible route to return an empty string (instead of declaring the function result to be ?string), so uplevel code can rely on always getting a string result.